refactor(network): share wifi profile xml generation
